new method
In this example, we're using a MyDatabase class to create and manage the Room database instance. The lastUpdated timestamp is stored as a timestamp field in the MyData table. When we fetch the data from the API, we include the If-Modified-Since header with the last updated timestamp, and the API returns the data only if it has been updated since that time. If there is new data, we insert it into the local Room database and update the last updated timestamp.